fix SIMDe emulation builds on Arm, add native translation from x86 for comparison
