+++ cmake/archdetect.cmake
@@ -67,7 +67,10 @@ if (USE_CPU_NATIVE)
         message(STATUS "clang will tune for ${GNUCC_ARCH}, ${TUNE_FLAG}")
     endif()
 else()
-    if (ARCH_IA32 OR ARCH_X86_64)
+    if (SIMDE_BACKEND)
+        set(GNUCC_ARCH native)
+        set(TUNE_FLAG generic)
+    elseif (ARCH_IA32 OR ARCH_X86_64)
         set(GNUCC_ARCH native)
         set(TUNE_FLAG generic)
     elseif(ARCH_AARCH64)
@@ -85,8 +88,8 @@ else()
        set(GNUCC_ARCH armv7a)
        set(TUNE_FLAG generic)
     elseif(ARCH_PPC64EL)
-       set(GNUCC_ARCH power9)
-       set(TUNE_FLAG power9)
+       set(GNUCC_ARCH power8)
+       set(TUNE_FLAG power8)
     else()
        set(GNUCC_ARCH native)
        set(TUNE_FLAG native)
