+++ cmake/simde.cmake
@@ -1,5 +1,9 @@
-include_directories(${PROJECT_SOURCE_DIR}/simde/simde)
+# include_directories(${PROJECT_SOURCE_DIR}/simde/simde)
 
-set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DSIMDE_BACKEND")
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DSIMDE_BACKEND")
+set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DVS_SIMDE_BACKEND")
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DVS_SIMDE_BACKEND")
 
+if (SIMDE_NATIVE)
+  set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DVS_SIMDE_NATIVE -DSIMDE_ENABLE_OPENMP -fopenmp-simd")
+  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DVS_SIMDE_NATIVE -DSIMDE_ENABLE_OPENMP -fopenmp-simd")
+endif()
